Switch temperature storage to float32 for precise diffusion
Grid._temp is now float32 (was int16); diffuse_temps drops the np.rint-to-int16 cast and returns float32. The int16 round-to-nearest was stalling cells just above phase-transition thresholds -- e.g. water cooling next to ice stuck at ~+6C and never reaching <=0 to freeze. With float precision, a cell cooling toward 0 actually crosses it (validated: crosses <=0 where int16 stalled at +2). Conservation drain drops from +/-15 (int16 rounding) to ~4e-5. Stale int16 comments in config/elements/simulation corrected.

+++ tests/test_grid.py
@@ -220,7 +220,7 @@ def test_temp_array_defaults_to_ambient() -> None:
 
     grid = Grid(width=4, height=4)
     assert grid.temp.shape == (4, 4)
-    assert grid.temp.dtype == np.int16
+    assert grid.temp.dtype == np.float32
     for y in range(grid.height):
         for x in range(grid.width):
             assert grid.get_temp(x, y) == AMBIENT_TEMP
